See description
Started navigation to compass screen (from friends list screen), updated compass screen logic (to use actual friend's location), and moved some compass screen variables and functions to user profile file
