+++ ProxiPalTest_client_20231205232005/app/src/main/java/com/mongodb/app/navigation/NavigationRoutes.kt
@@ -15,7 +15,7 @@ sealed class Routes(val route: String) {
     object OnboardingScreen : Routes("onboardingscreen")
     object AdvancedScreenSettings : Routes("advancedscreensettings")
     object MessagesScreen: Routes("messagesscreen")
-
+    object CompassScreen: Routes("compassscreen")
     object FriendListScreen : Routes("friendlist")
     object FriendRequestScreen : Routes("friendrequestscreen")
 }
